Move stored payment sources to User graphql object
